Fix cursor location after replacement
